bot_bs_helper/requirements.txt
aiogram==2.14.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
babel==2.16.0
bluesales-python-sdk==1.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
progress==1.6
propcache==0.2.1
pytz==2024.2
requests==2.32.3
urllib3==2.3.0
vk-api==11.9.9
yarl==1.18.3
